Miscellaneous tensor testing improvements
Includes:
6D tensor shapes.
MSVC warning fixes.
Upper case format names in test names.
Tensor memory barrier refactor.
More complete tensor shape support checks.
Workgroup sizes increased to keep group counts below Vulkan minimum limit.

Affects:
dEQP-VK.tensor.*

Components: Vulkan

VK-GL-CTS issue: 6895

+++ external/vulkancts/framework/vulkan/vkTensorMemoryUtil.cpp
@@ -28,15 +28,15 @@
 namespace vk
 {
 
-const TensorStrides getTensorStrides(const TensorDimensions &dimensions, size_t formatSize, size_t multiplier)
+const TensorStrides getTensorStrides(const TensorDimensions &dimensions, size_t formatSize, int64_t multiplier)
 {
     DE_ASSERT(dimensions.size() > 0);
     TensorStrides strides{};
 
     for (size_t dim_idx = 0; dim_idx < (dimensions.size() - 1); ++dim_idx)
     {
-        strides.push_back(multiplier * std::accumulate(dimensions.cbegin() + dim_idx + 1, dimensions.cend(), formatSize,
-                                                       std::multiplies<size_t>()));
+        strides.push_back(multiplier * std::accumulate(dimensions.cbegin() + dim_idx + 1, dimensions.cend(),
+                                                       static_cast<int64_t>(formatSize), std::multiplies<int64_t>()));
     }
 
     strides.push_back(formatSize);
